More from plugin into core and fix rename errors
