+++ glue/core/link_helpers.py
@@ -9,7 +9,8 @@ multiple ComponentLinks easily. They are meant to be passed to
 
 import types
 
-from glue.config import link_function
+from glue.config import link_function, link_helper
+
 from glue.core.data import ComponentID
 from glue.core.component_link import ComponentLink
 
@@ -17,7 +18,8 @@ from inspect import getfullargspec
 
 
 __all__ = ['LinkCollection', 'LinkSame', 'LinkTwoWay', 'MultiLink',
-           'LinkAligned', 'BaseMultiLink', 'ManualLinkCollection']
+           'LinkAligned', 'BaseMultiLink', 'ManualLinkCollection',
+           'JoinLink']
 
 
 @link_function("Link conceptually identical components",
@@ -463,3 +465,53 @@ def functional_link_collection(function, labels1=None, labels2=None,
     FunctionalLinkCollection.description = description or ''
 
     return FunctionalLinkCollection
+
+@link_helper(category="Join")
+class JoinLink(LinkCollection):
+    cid_independent = False
+
+    display = "Join on ID"
+    description = "Join two datasets on a common ID. Other links \
+in glue connect data columns (two datasets have 'age' columns but \
+the rows are different objects), while Join on ID connects the same \
+rows/items across two datasets."
+
+    labels1 = ["Identifier in dataset 1"]
+    labels2 = ["Identifier in dataset 2"]
+
+    def __init__(self, *args, cids1=None, cids2=None, data1=None, data2=None):
+        # only support linking by one value now, even though link_by_value supports multiple
+        assert len(cids1) == 1
+        assert len(cids2) == 1
+
+        self.data1 = data1
+        self.data2 = data2
+        self.cids1 = cids1
+        self.cids2 = cids2
+
+        self._links = []
+
+    def __str__(self):
+        # The >< here is one symbol for a database join
+        return '%s >< %s' % (self.cids1, self.cids2)
+
+    def __repr__(self):
+        return "<JoinLink: %s>" % self
+
+    # Define __eq__ and __ne__ to facilitate removing
+    # these kinds of links from the link_manager
+    def __eq__(self, other):
+        if not isinstance(other, JoinLink):
+            return False
+        same = ((self.data1 == other.data1) and
+                (self.data2 == other.data2) and
+                (self.cids1 == other.cids1) and
+                (self.cids2 == other.cids2))
+        flip = ((self.data1 == other.data2) and
+                (self.data2 == other.data1) and
+                (self.cids1 == other.cids2) and
+                (self.cids2 == other.cids1))
+        return same or flip
+
+    def __ne__(self, other):
+        return not self.__eq__(other)
