+++ glue/dialogs/link_editor/qt/data_graph.py
@@ -21,11 +21,11 @@ def get_pen(color, linewidth=1, linestyle=Qt.SolidLine):
 
 class Edge(QGraphicsLineItem):
 
-    def __init__(self, node_source, node_dest, linewidth=3, zindex=5, JoinLink=False):
+    def __init__(self, node_source, node_dest, linewidth=3, zindex=5, join_link=False):
         self.linewidth = linewidth
         self.node_source = node_source
         self.node_dest = node_dest
-        if JoinLink:
+        if join_link:
             self.linestyle = Qt.DashLine
         else:
             self.linestyle = Qt.SolidLine
@@ -166,12 +166,12 @@ def get_connections(dc_links):
     for link in dc_links:
         data1 = link.data1
         data2 = link.data2
-        if link.JoinLink:
-            JoinLink = True
+        if link.join_link:
+            join_link = True
         else:
-            JoinLink = False
+            join_link = False
         if (data1, data2) not in links and (data2, data1) not in links:
-            links.append((data1, data2, JoinLink))
+            links.append((data1, data2, join_link))
 
     return links
 
@@ -284,14 +284,14 @@ class DataGraphWidget(QGraphicsView):
         # Get links and set up edges
 
         if old_links:
-            self.background_edges = [Edge(self.data_to_nodes[data1], self.data_to_nodes[data2], linewidth=1, zindex=1, JoinLink=JoinLink)
-                                     for data1, data2, JoinLink in get_connections(data_collection.external_links)]
+            self.background_edges = [Edge(self.data_to_nodes[data1], self.data_to_nodes[data2], linewidth=1, zindex=1, join_link=join_link)
+                                     for data1, data2, join_link in get_connections(data_collection.external_links)]
         else:
             self.background_edges = []
 
         if new_links:
-            self.edges = [Edge(self.data_to_nodes[data1], self.data_to_nodes[data2], JoinLink=JoinLink)
-                          for data1, data2, JoinLink in get_connections(new_links)]
+            self.edges = [Edge(self.data_to_nodes[data1], self.data_to_nodes[data2], join_link=join_link)
+                          for data1, data2, join_link in get_connections(new_links)]
         else:
             self.edges = []
 
@@ -317,8 +317,8 @@ class DataGraphWidget(QGraphicsView):
         for edge in self.edges:
             edge.remove_from_scene(self.scene)
 
-        self.edges = [Edge(self.data_to_nodes[data1], self.data_to_nodes[data2], JoinLink=JoinLink)
-                      for data1, data2, JoinLink in get_connections(links)]
+        self.edges = [Edge(self.data_to_nodes[data1], self.data_to_nodes[data2], join_link=join_link)
+                      for data1, data2, join_link in get_connections(links)]
 
         for edge in self.edges:
             edge.update_position()
