--- glue/plugins/join_on_key/__init__.py
@@ -1,2 +1,0 @@
-def setup():
-    from . import link_helpers  # noqa
